requirements.txt
asgiref==3.6.0
Babel==2.12.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
django-money==3.0.0
djangorestframework==3.14.0
idna==3.4
psycopg2-binary==2.9.5
py-moneyed==2.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
